+++ src/modules/pages/Accessories/AccessoriesPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import CategoryPage from '../../shared/components/CategoryPage';
+import { getAccessories } from '../../../utils/fetchClient';
+
+export const AccessoriesPage: React.FC = () => {
+  return <CategoryPage type={'accessories'} getProducts={getAccessories} />;
+};
+
+export default AccessoriesPage;
