+++ src/modules/pages/Tablets/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TabletsPage';
