+++ src/modules/shared/components/CategoryPage/CategoryPage.module.scss
@@ -1,4 +1,4 @@
-@use '../../../utils/mixins';
+@use '../../../../utils/mixins';
 
 .icons {
   display: flex;
@@ -24,9 +24,4 @@ a {
   letter-spacing: 0%;
   color: #89939A;
   margin-bottom: 40px;
-}
-
-.form {
-  display: flex;
-  gap: 16px;
 }
