+++ src/modules/shared/components/ProductsList/ProductsList.module.scss
@@ -25,3 +25,8 @@
     }
   }
 }
+
+.form {
+  display: flex;
+  gap: 16px;
+}
